requirements.txt
keras==2.10.0
Keras-Preprocessing==1.1.2
matplotlib==3.6.1
numpy==1.23.4
opencv-python==4.6.0.66
pandas==1.5.1
Pillow==9.3.0
scikit-learn==1.1.3
streamlit==1.14.0
tensorflow==2.10.0
